#24 int를 Integer와 매칭되도록 코드 변경

+++ codebone-generator/src/main/java/org/codebone/generator/connector/Column.java
@@ -25,7 +25,7 @@ public class Column {
 		if ("int".equals(type) && DatabaseType.MSSQL == databaseType) {
 			transformedType = "Integer";
 		} else if ("int".equals(type) && DatabaseType.MYSQL == databaseType) {
-			transformedType = "Long";
+			transformedType = "Integer";
 		} else if ("tinyint".equals(type)) {
 			transformedType = "Integer";
 		} else if ("float".equals(type)) {
